requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
psycopg2
psycopg2-binary
django-cors-headers